Boost Rings overhaul
alright, made the edits to the rings. They're MOSTLY deterministic now, so the vast majority of the physics isn't run by a timeline or anything, so if you interrupt it it'll work mostly as normal.
They're also way more consistent and I think are easier to set up.
The player now consitently launches from the ring, to the 3D vector offset. The arrow still doesn't update right but I think that's out of my control.
As soon as the player would reach the end of that arrow, normal physics and controls return.
There is now an additional float for them, that controls how fast the player is launched from the ring to the location. A speed of 1 takes one second, a speed of 2 takes 0.5 seconds.
but again it's not a lerp, so a higher speed will mean the physics applied to them when they pass the line will be speedier too.
imo it's more intuitive than I'm making it sound.
Your current ones probably need their values fixed... again... Sorry, I'll try to make this the last that happens.
Oh also, rotations are set up, those actually do use a timeline. Animation speed is effected by boost speed too, so a slower boost has a slower spin.
Code is all commented now too if you care to look or make edits.